requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-dotenv
